Move tests to particular files
